+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v0.74.26, 1 November 2018
+
+- Handle case with no previous dependabot PRs
+
+## v0.74.25, 1 November 2018
+
+- Use `chore` as semantic commit message if it is being used
+
 ## v0.74.24, 1 November 2018
 
 - Yarn: Fix resolved url in lockfile for dependencies installed from GitHub using the